changed test to be `final byte` rather than `final int` to be consistent with total amount of bits available in the capabilities mask.

+++ src/test/java/org/reaktivity/specification/nukleus/CoreFunctionsTest.java
@@ -102,7 +102,7 @@ public class CoreFunctionsTest
     {
         final byte[] challengeMaskBytes = CoreFunctions.capabilities("CHALLENGE");
         ByteBuffer bb = ByteBuffer.wrap(challengeMaskBytes);
-        final int challengeMask = bb.get();
+        final byte challengeMask = bb.get();
         assertEquals(0x01, challengeMask);
     }
 }
